projecthlulihlawu/requirements.txt
altgraph==0.17.2
asgiref==3.5.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
docutils==0.19
exceptiongroup==1.0.0rc9
future==0.18.2
futures==3.1.1
goslate==1.5.2
h11==0.14.0
idna==3.3
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
mock==2.0.0
MouseInfo==0.1.3
oauthlib==3.2.0
outcome==1.2.0
pbr==5.9.0
pefile==2021.9.3
phonenumbers==8.12.57
Pillow==9.2.0
plyer==2.0.0
PyAutoGUI==0.9.53
pycparser==2.21
PyDictionary==2.0.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pypiwin32==223
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pyttsx3==2.90
pytweening==1.0.4
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.10.0
tzdata==2022.1
urllib3==1.26.9
Vocabulary==1.0.4
wikipedia==1.4.0
win10toast==0.9
wsproto==1.2.0asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==4.4.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
django-unicorn==0.49.0
django-widget-tweaks==1.4.12
forex-python==1.8
idna==3.4
orjson==3.8.1
phonenumbers==8.13.0
Pillow==9.3.0
requests==2.28.1
shortuuid==1.0.11
simplejson==3.18.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.1.1
CurrencyConverter==0.17.4
decorator==4.4.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
django-unicorn==0.49.0
django-widget-tweaks==1.4.12
forex-python==1.8
idna==3.4
orjson==3.8.1
phonenumbers==8.13.0
Pillow==9.3.0
requests==2.28.1
shortuuid==1.0.11
simplejson==3.18.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
